Copying application from different repo
This version of the application is mostly functional. There are a few buttons not hooked up but the server and client can talk to each other and data/models/forecasts are plotted successfully.

Currently its setup to run locally. Prophet can be a bit of a pain to install via requirements.txt file.